ch: fix failing tests
 - fix failing tests

[Maintains #172279704]

+++ backend/settings.py
@@ -199,7 +199,7 @@ EMAIL_USE_TLS = True
 DEFAULT_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
 EMAIL_BACKEND = 'backend.apps.email.backends.django_q.DjangoQBackend'
 DJANGO_Q_EMAIL_BACKEND = config('DJANGO_EMAIL_BACKEND', DEFAULT_BACKEND)
-VERIFY_EMAIL_LINK_AGE = config('VERIFY_EMAIL_LINK_AGE', 8 * 60 * 60) # 8-days
+VERIFY_EMAIL_LINK_AGE = int(config('VERIFY_EMAIL_LINK_AGE', 8 * 60 * 60))  # 8-days
 
 # Django-q configs
 Q_CLUSTER = {
